Adds SCEP extension client
