+++ backend/pkg/manufacturing/client/client.go
@@ -1,12 +1,13 @@
 package client
 
 import (
+	"context"
 	"crypto"
 	"crypto/tls"
 	"crypto/x509"
 )
 
 type Client interface {
-	StartRemoteClient(CA string, authCRT []tls.Certificate) error
-	GetCertificate(keyAlg string, keySize int, c string, st string, l string, o string, ou string, cn string, email string) (*x509.Certificate, crypto.PrivateKey, error)
+	StartClient(ctx context.Context, CA string, authCRT []tls.Certificate) error
+	GetCertificate(ctx context.Context, keyAlg string, keySize int, c string, st string, l string, o string, ou string, cn string, email string) (*x509.Certificate, crypto.PrivateKey, error)
 }
