+++ backend/pkg/manufacturing/client/extension/testdata/revoked.crt
@@ -1,0 +1,33 @@
+-----BEGIN CERTIFICATE-----
+MIIFtzCCA5+gAwIBAgIBBDANBgkqhkiG9w0BAQsFADBjMQswCQYDVQQGEwJFUzER
+MA8GA1UECAwIR2lwdXprb2ExETAPBgNVBAcMCEFycmFzYXRlMRswGQYDVQQKDBJM
+S1MgTmV4dCwgUy4gQ29vcC4xETAPBgNVBAMMCGVucm9sbGVyMB4XDTIwMTEyNjEx
+NDY0OVoXDTIxMTEyNjExNDY0OVowYzELMAkGA1UEBhMCRVMxETAPBgNVBAgTCEdp
+cHV6a29hMREwDwYDVQQHEwhBcnJhc2F0ZTEbMBkGA1UEChMSTEtTIE5leHQsIFMu
+IENvb3AuMREwDwYDVQQDEwhvcGVyYXRvcjCCAiIwDQYJKoZIhvcNAQEBBQADggIP
+ADCCAgoCggIBALy4/1SetbjBYiXlkbmiG6MwR421U7M5YwFB5ckbKFt7lqa+FHxc
+v1obeYJ1dx5eAhbleZJgYyFD4ruQ9N4ZeHSztai0NK3dnpoq31hkPwEMRsDncQDV
+NlgOzNfI/0bnkB549zb3VNps7doySEhoTmVivf8pvkIbDFUyeUkIbN1Yy7dmUmVE
+WFi8dK5q0JwKRKrjmDDT+8wgCN6lcM6kS6CoXH8xEwCrIYz4U3ktI1hEb7Dc0mAF
+cx574nkmvDeau3MxwvTbU9YpVr2YaR32LBQvX0VvHO7zwY7Sh009+oCdB1zHcgHB
+PGkeKuJ4cUxDK7VTBB1YGZnOv508z0jLL2rF1LV4I7fVpnTy4q+WsbmIVXA3d64Z
+lqqmmSLK4FCSi+SgsSp2gkBUtp8VJqyxSSTBtX6704paoW1/qPicJQ19hwnjAuPE
+CASrIo1bCrWs4hIf7N+Yy/NELfwn9oSxdpfm94tdBURfZ9Zny71G7h3ftmmDr6Sv
+1Jsm2bqmJH72RMHaaiMYZxfyiBFCIAvYa5XPeVaxnw5GSredb0nP9wrERjJPn7hB
+0DYchWiyWsztSb6PNQjkyPu83e7dKEKfjZiYfibmzsygVpMPr/bKRYii6fEQSqyD
+IZUUO8lWEu7Pg1R4GOsazpg4QtpvQ18+G+XcdxfZ/jijmxKdNh8Nij6HAgMBAAGj
+djB0MA4GA1UdDwEB/wQEAwIHgDATBgNVHSUEDDAKBggrBgEFBQcDAjAfBgNVHSME
+GDAWgBRxz+zHTXe3Jhaxd4kaAKQK+qlz4zAsBggrBgEFBQcBAQQgMB4wHAYIKwYB
+BQUHMAGGEGh0dHA6Ly9vY3NwOjkwOTgwDQYJKoZIhvcNAQELBQADggIBAMEWw0Jw
+W20IjHlxAQ6KrxvIOwE2OvCmi6mr3+8eZNCkMGrCqKYxtgSjdncRIlKfI3y7wqd3
+v38qyRk6S+pxTgBVEqwD/PPJvBZats5cWek2EhK8nYHz1Olfk1AK6XKwG57PUE2c
+787+3Bmtfl5hZ1FXmQRaoQxj/k+tHEwI8ezaHw2pfgX5rLxtUMtMpVGNaTMhSKmu
+rvi3Tht2GiEejqlR52lZ0ajU63iX2GCSrPM6W3pHqyU5zFCW7uA27d68ePiS3qkh
+0966J4EmPP6Tdt3BjYzH4bvpoLciubA6LSl/jwYhUTd8G+RLIHy6qBUSFGJ+YDU8
+/ZFsLiZyi9dzOlNNkRxXO67VcIfI76RH+wxui+BMF+wksRpka4bb4hvkzwtazh/J
+OBCs5NmsknjJu1VofcOPzWHKXgkg4hQRMCq87lWjYbqf5cm7gfdW0JmISjD5wGXs
+og/aIUwRKaw9D3QUzuVhxdv9sGS+oyQX93DU/YEWm/RO9adebJu3QkchbzC2nAl4
+2ly1LnqpvK/kkacaSvGFNZ08CkvuEuFbdoZZ9mPsUBXZhedf0KZUNyw5Dft07AHV
+hgng30VM4PqoRhEiWmYzMBQ/kLdUuO0cdfuoF9Dm8OcSwrO2HNB4vNF3QMvZYDBu
+pf9mHiob2/71hsXWrKmwkkK+lfrhXB1HA/Np
+-----END CERTIFICATE-----
