+++ backend/pkg/manufacturing/client/extension/testdata/self.crt
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDnzCCAoegAwIBAgIUU1GQ3tWZhguCu7M4hHOhwgrkW2wwDQYJKoZIhvcNAQEL
+BQAwXzELMAkGA1UEBhMCRVMxETAPBgNVBAgMCEdpcHV6a29hMREwDwYDVQQHDAhB
+cnJhc2F0ZTEbMBkGA1UECgwSTEtTIE5leHQsIFMuIENvb3AuMQ0wCwYDVQQDDAR0
+ZXN0MB4XDTIwMTEyNjExNTAwNloXDTIwMTIyNjExNTAwNlowXzELMAkGA1UEBhMC
+RVMxETAPBgNVBAgMCEdpcHV6a29hMREwDwYDVQQHDAhBcnJhc2F0ZTEbMBkGA1UE
+CgwSTEtTIE5leHQsIFMuIENvb3AuMQ0wCwYDVQQDDAR0ZXN0MIIBIjANBgkqhkiG
+9w0BAQEFAAOCAQ8AMIIBCgKCAQEA1jkxlKSUCLYNu+RXUqe/X91s6AiUoCJBb3pL
+bvAX7La9fDe2v6u+aqyQMB6CzDlmo2BUib0t0oo/77ygZdEdbzkhTgppifWd/SFg
+/V5nZm3fmiRWgsJbwYMtQdKTfVsTz20aYLNWhh8k+4zh6QWt9QRL5EPVK0oeJpoN
+vRd/ssakTw5n0cpTSEOYIJD8QxXZ03ItwZOf+53xCqrbQH/XCZkRR48roZmPXMaa
+ff3nr2lDpM8VUnmu0sjOVz5aRfaifPFrzmdm2jWrNy5pmAyu1Jjd/B+a0xhlXhBR
+Z2HttPMnDJFHVs2kZRNKM3t4JAZVra37eEpV9esDN/cQX/Vp2QIDAQABo1MwUTAd
+BgNVHQ4EFgQU6uAhyp7IxlUhzrHbKd2M193bSBAwHwYDVR0jBBgwFoAU6uAhyp7I
+xlUhzrHbKd2M193bSBAwDwYDVR0TAQH/BAUwAwEB/zANBgkqhkiG9w0BAQsFAAOC
+AQEAYEqPF2LTGt8iGKKNwJ6pC3gzaiXtoFePn0XdRgqJKxlah+aed3ZvITPhnyhW
+VUePvnpYj1TNqYJZil2q75zo3a4QEU/4rttfrSmmIKXE6yFmXoKLwiTMmk14KCaR
+EIrglZ2jiNesvdYKrSP7f4jSKE+8atb0Qot3w4a0cD5FVj1ZStxo+tfeylFPyisg
+AWBs3BVRNUdCf/3LCw1LC52rrchvXyFVnzBenWQ5z6zlZVTNO1cl/G8U/cNVlJ00
+5dVSzbkWIj7Jhb2VQYtCcmFMhsiJBHlyEmGhrIdvLFYXRGb93HXQFXfFuoIQbLUJ
+eOOSH1OJUmhOW7wpmm7WpLHjSg==
+-----END CERTIFICATE-----
