+++ backend/pkg/manufacturing/client/extension/testdata/valid.key
@@ -1,0 +1,51 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIJKQIBAAKCAgEAvLj/VJ61uMFiJeWRuaIbozBHjbVTszljAUHlyRsoW3uWpr4U
+fFy/Wht5gnV3Hl4CFuV5kmBjIUPiu5D03hl4dLO1qLQ0rd2emirfWGQ/AQxGwOdx
+ANU2WA7M18j/RueQHnj3NvdU2mzt2jJISGhOZWK9/ym+QhsMVTJ5SQhs3VjLt2ZS
+ZURYWLx0rmrQnApEquOYMNP7zCAI3qVwzqRLoKhcfzETAKshjPhTeS0jWERvsNzS
+YAVzHnvieSa8N5q7czHC9NtT1ilWvZhpHfYsFC9fRW8c7vPBjtKHTT36gJ0HXMdy
+AcE8aR4q4nhxTEMrtVMEHVgZmc6/nTzPSMsvasXUtXgjt9WmdPLir5axuYhVcDd3
+rhmWqqaZIsrgUJKL5KCxKnaCQFS2nxUmrLFJJMG1frvTilqhbX+o+JwlDX2HCeMC
+48QIBKsijVsKtaziEh/s35jL80Qt/Cf2hLF2l+b3i10FRF9n1mfLvUbuHd+2aYOv
+pK/UmybZuqYkfvZEwdpqIxhnF/KIEUIgC9hrlc95VrGfDkZKt51vSc/3CsRGMk+f
+uEHQNhyFaLJazO1Jvo81COTI+7zd7t0oQp+NmJh+JubOzKBWkw+v9spFiKLp8RBK
+rIMhlRQ7yVYS7s+DVHgY6xrOmDhC2m9DXz4b5dx3F9n+OKObEp02Hw2KPocCAwEA
+AQKCAgEAqCuyW5WSGbE55MtNsD3koubC3GRY0byxnM08Fpk8vUjEVpHLb4a2fXU+
+MiK6ozTiZ8TVMLrzmxEDWmEx0zifPl9l4Y+z4bs7OeY6YGn0OtrVsdCLwlI8uvt1
+TlSuJmDy7DSfFzwPQRUQXv+vqdZJBc+Xv2lTQVUZCFKSAaKgYB5gam2jvtnIYEqP
+L12/x0zfI+jsFZbRKnThcoW6oGmHewKMaksww8e7fgNLXw/6Dw3s+Y3NagNxMhjz
+F0KNNn+LynrtX4SSkKUFxq2p/ij/3nzYSOdawRQ1xy+ToiE4uKM1jFeuICIx0vou
+/OCXFX5F1K7FW+DF72ltFpC5pgRpmxOu35Im++F35frkZIh6nH136OTRmjtk17Tf
+Tehf4n6Tx5kYinVQatAebCVQy/Ms4+JP4H6CF8UwyBq5Lesr+6FtkI91ODOjhgs7
+lh8isYtiZzzahHuUZ032t0jXofM3orIze1pqynlKJyTSEWsGL9JLMHzXd7c3ovDI
+BVij0tAE6IZO9zMfc+evWbSKyMT5eF7t6QA37GT+5/4BaJe8K+WUJhP8Zat7wZYG
+Ur2QHZhCe3VAw9bNkrUEM/wYRjpDnJVBwkslCHZm/nAxEy/NCcM8qdRhLs2VwRZ9
+sTC/Ud2c0vBHHMyZN62ThP05ssWK8aNKQHra672a1UFPsZnV1QECggEBAOdZbv4x
++qwtr43JPgXIgBb0jWlE9NMda0UIlPGrNWnYQmissdMjZ5829BYiviEWc+4kXTlN
+a/6SpDyYuLZoikUrKxiUiXjkCj3gAAzQ0nRhnOmnJiZ1Fo3w6CikjiZZ+P2g8pNu
+2BBVaCmnGxrU7kLNspccgP8kLuKtP/LOXwIo8qdwZq81K0txe/SBx7zDOeDLmyJ7
+sJClHxPDOLhtMmhtY79xy9x0vqpgLVFHqZPWYiz1sCBStlX6FJZHBhnNUfq+l7Fq
+i+dMKvfT2+Gi2GdlFXA+L76SNYpo2/xxzaDrjerS2Xp7Sy0crN2877anppPJSEQo
+hLKf/PClce09VXsCggEBANDU03J7JVk37o2X1T6S7yHk2xrP1bGloXXNs4xqyItb
+X9AS52TUrRTYMnwDS4wpVOFeDxInMhs2tLtFy98rlGpBj2iigETY0a7CwaahVi8E
+F+7dfrcaz+/CdKXmd0RxHCINUHxFCvDblOSJrGzNLhpV+cyynGbWdngq1tuOCz8N
+HQCDot801uEB0Qr36bPYWC2iVfBYDcRTeoheU3i6OUfJmy/wyrIUthIF6ZUamAVJ
+MTczLFZpLxZFuXOS0CZSV1cOtgIE67yeuaGKc2ry9A3C8ScVl1rp7QVpIHygIuD0
+/DgnSqGhsgLS6gX1Un7oSt0/QiqvhPt499AIYyj2/2UCggEBALDDmpTKZzeLKKly
+69rLWWMws6ONNmrf/YgfiH4mTId03zjPSoeCc7go6tLG9PKfDzETw872ot89+BL8
+tZVZxtwWobUEOQ2PX2vrbN5d549YzeWQ+wtDR1poiVoUcTBIaoCPqwiPFrNULmQ0
+8/bKXsnYFO50rvNDGDfGHChPGjSySMU6kpqoS043hXhKeOx3Vg+2QfBVBZFfOKGG
+9h2kAXHcsuwh7Jui79OJ7NY/pQf6M8Uaa9ommH6Ut4oi9jxcrpiha5H5ULAHeIUm
+8Mgz/pjxmTHLsfswB4yblgoKggzELgkP1W9KKSRs1oANzjKHYjVlA7xgIRPCPic4
+ohWcWWUCggEAAhCYpqEhFWjAKxY9PYkmZCdnIzS64C7LEecex1ym7vnQULZfI1gW
+/gXFOgSuUox2EmRjscJBdJNt5gwin+pGZWaQjHBO5F9gc2HpmEQwHfbkZbA6+ISD
+WdOKY7mLpFmZ9Zw8F353RTpmlnvmAI1mz0vaVHPPqBSmxzllNnaJFe6DQuYKTjqv
+hToHXxxkayrQdGMOhnnYaQCKTZ7hJGhy+Iq9xhOTbX9W7YaxcI/6wNKgcfNHlPMl
+TAsUclgZRse5pVq34f06d3r6SuabjYspbplm2QQVowh9F3iaoaDTt3W0qLl6u25l
+3DFJrxVl42tYkid1+AYUQohHqv7qPBV1IQKCAQAzGMd7ppSeaKWximCyZUFDd+PW
+4PsS3dgv5xOJvLSESvZ4zYy07yLnERBbVqguCBK0FI+60r6hGwDjYbQTF8oI00Xw
+2DvB6O0BntyEzHnrcAlcLb2iRhTCM+iklHp0JJVmbioijglCqMD09z1wK0QK0yaY
+ltFfCynLsFU9q90HPnFBnSYDIVCx+/MF+i87S5JTayJPeaFC+kiIQvYgHaGwBbAt
+UG1UlMjS/rfzbKlGqfjxIMzcTeqOYMDtVgeXC+BFhjrJDynmGo1ELUZ/B9Gl+0cq
+Itva3xJoDtZIZggJJnvOfLMfUUk08kbm6rN0OfupCeIxKc7wqIzXUWmBSfon
+-----END RSA PRIVATE KEY-----
