--- backend/pkg/manufacturing/client/scep/testdata/self.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEowIBAAKCAQEA1jkxlKSUCLYNu+RXUqe/X91s6AiUoCJBb3pLbvAX7La9fDe2
-v6u+aqyQMB6CzDlmo2BUib0t0oo/77ygZdEdbzkhTgppifWd/SFg/V5nZm3fmiRW
-gsJbwYMtQdKTfVsTz20aYLNWhh8k+4zh6QWt9QRL5EPVK0oeJpoNvRd/ssakTw5n
-0cpTSEOYIJD8QxXZ03ItwZOf+53xCqrbQH/XCZkRR48roZmPXMaaff3nr2lDpM8V
-Unmu0sjOVz5aRfaifPFrzmdm2jWrNy5pmAyu1Jjd/B+a0xhlXhBRZ2HttPMnDJFH
-Vs2kZRNKM3t4JAZVra37eEpV9esDN/cQX/Vp2QIDAQABAoIBAHS12j5KX+BTFy5r
-m+uLOqdG4oOT0S1hHFdYwGojYlVzxc48KFNKhUZm13Nv3ADZsQKHg+c66xEOgwDg
-ul5eTHQBIxT6dT9VUorLOaUwXpMefzjiXUOGEilehzSaPgKIsUVMApf2fMg7hZGK
-0rV6g6i+e35UOnoIOl77hUqk18yUjohHygv7ZlZIirS+tOCSAgmpEh776ie0+/U4
-yV91/SudSBhZ/dSdTPTOH574D0181ykPAw63l8MBk/14/EBLnWshz5K2mCf1ZfDM
-u9z07Fo0ebTrydpqgsN6eljWiquq403pxZg8IDEUeT8Gkr/VzLWd2utkibjHA/Zd
-L0ZqxJUCgYEA/befb5pRzzKrslgxOxEcsoEstO7oU6vYarNwrRTCNP4Qu38y97iY
-KCt94EW7zAUDp8cvdJFT5L+oKTUjicpd3M/9DdMY8iwFSQ+f2kYhIkoh5uwNHnMC
-ll7ng8W84W0YiPftejtNT2fF2IvBXGrRFIUNQQEvOLpzMaKpnZz0+lcCgYEA2Cab
-Lw2ufQgOuUx/UbHVkocOHXFPLb31AYn94nvnh3ZIC4UT2bHRbr+w/RBRNLLoESpY
-1suybhLqys98kCjxi1fpQYVNu4lGSb8zq34cCVwxDz5vqOuUV09snfiYLnyR6k/i
-RpfzXLs/u5vXNIkSom+6MvXQzg2gNVliB0RNf08CgYA+YhHPrRAEdB8HYOvp7BE0
-StKlQkDZod06jxhbVklSIk4W6/WqNzciWV8/76V3SpxVeoJOlRUdn4pHZ4tva9Sc
-YH+GEKuHkA9UoYTadHfdKIEheFTZccnnEzq1ib8hZmOr+dnoPgCKT6ez0NkgUhtK
-glR58UKdHisIv0gZOuh5TQKBgDxsDlVI7c3lwrfpum2L2LftKPTyB8AxqZuUh3QY
-TtB/MASIXtQr8O+UhlU3invAxz6FJ/BjYK2EfD3VYmyqBVt+Z6X5LZGO/ib/W+JE
-jnHPY6dvPKdrszv059gA9gmIqQFm80nokA8MCPTYy3/UEXF26gX00TnIdVB0fAKs
-BT3hAoGBANiPosdZACt9Yn/P8Xs4IAgns/541A87mOVxiTEyUdxFa7oxO9QHfmKL
-XdXwWYzhpNqmF6By3sSniHBwwJwFFIj7aWjcaxJkAp22BrP51sGMUWMPTX1eY6zl
-0iqafCpcNgVYUDiMZsZl8ydk8AZp2Cd32nr+BXyT7fJM8ypi/J75
------END RSA PRIVATE KEY-----
