--- backend/pkg/manufacturing/client/scep/testdata/valid.crt
@@ -1,33 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFtzCCA5+gAwIBAgIBAzANBgkqhkiG9w0BAQsFADBjMQswCQYDVQQGEwJFUzER
-MA8GA1UECAwIR2lwdXprb2ExETAPBgNVBAcMCEFycmFzYXRlMRswGQYDVQQKDBJM
-S1MgTmV4dCwgUy4gQ29vcC4xETAPBgNVBAMMCGVucm9sbGVyMB4XDTIwMTEyMDEw
-MTUwMloXDTIxMTEyMDEwMTUwMlowYzELMAkGA1UEBhMCRVMxETAPBgNVBAgTCEdp
-cHV6a29hMREwDwYDVQQHEwhBcnJhc2F0ZTEbMBkGA1UEChMSTEtTIE5leHQsIFMu
-IENvb3AuMREwDwYDVQQDEwhvcGVyYXRvcjCCAiIwDQYJKoZIhvcNAQEBBQADggIP
-ADCCAgoCggIBALy4/1SetbjBYiXlkbmiG6MwR421U7M5YwFB5ckbKFt7lqa+FHxc
-v1obeYJ1dx5eAhbleZJgYyFD4ruQ9N4ZeHSztai0NK3dnpoq31hkPwEMRsDncQDV
-NlgOzNfI/0bnkB549zb3VNps7doySEhoTmVivf8pvkIbDFUyeUkIbN1Yy7dmUmVE
-WFi8dK5q0JwKRKrjmDDT+8wgCN6lcM6kS6CoXH8xEwCrIYz4U3ktI1hEb7Dc0mAF
-cx574nkmvDeau3MxwvTbU9YpVr2YaR32LBQvX0VvHO7zwY7Sh009+oCdB1zHcgHB
-PGkeKuJ4cUxDK7VTBB1YGZnOv508z0jLL2rF1LV4I7fVpnTy4q+WsbmIVXA3d64Z
-lqqmmSLK4FCSi+SgsSp2gkBUtp8VJqyxSSTBtX6704paoW1/qPicJQ19hwnjAuPE
-CASrIo1bCrWs4hIf7N+Yy/NELfwn9oSxdpfm94tdBURfZ9Zny71G7h3ftmmDr6Sv
-1Jsm2bqmJH72RMHaaiMYZxfyiBFCIAvYa5XPeVaxnw5GSredb0nP9wrERjJPn7hB
-0DYchWiyWsztSb6PNQjkyPu83e7dKEKfjZiYfibmzsygVpMPr/bKRYii6fEQSqyD
-IZUUO8lWEu7Pg1R4GOsazpg4QtpvQ18+G+XcdxfZ/jijmxKdNh8Nij6HAgMBAAGj
-djB0MA4GA1UdDwEB/wQEAwIHgDATBgNVHSUEDDAKBggrBgEFBQcDAjAfBgNVHSME
-GDAWgBRxz+zHTXe3Jhaxd4kaAKQK+qlz4zAsBggrBgEFBQcBAQQgMB4wHAYIKwYB
-BQUHMAGGEGh0dHA6Ly9vY3NwOjkwOTgwDQYJKoZIhvcNAQELBQADggIBAEkaVhpw
-gYg8XZUz0U/JIpjyXgk/DIjScNHimoxhVpQky8SBG06avYnTN/D6SsB+4tByw3yI
-PfLHQPlOb/XtKAB8umrfVCsGmpYUVqNh9MSGfgYQAg38JvBcAElmPmxnl4sl373W
-zkJa/BEn9YW6UFVMjKvb3RA3MUr/7YzmtewIftweyCWX9oUaCekPEbxebMal6cRO
-B9OSCT38cRRM3jYBU6+x0dFY3ORChh3j/6UR13ZG8CH9DnHDUueTglVunpX3zP3j
-SQh964/GHpSnZXvJ2P5eyKAbvWFCs5XpbEGo0nMVUZFPNFMQZYpjQLIL7jBrEh9S
-oDkloMUU5mAcoqtkSBRdxFSh/+irZBWbpjjgR2nKIXSQvL153HZ3y6jZzw6QX5+g
-ef8Sqh8qAE8hGHillaeidQBwdFVYN+M8bpcaMmdcpxke+GU8kPamFIi3nQk1x/+y
-n9ToBVr43StrB/DjeIPTS5E9fbKY6Xqdi2sk+p4XvJkQAi8TcEnNq/jRR47U/PSl
-9QXFPd/tx9aFsprE986kSNrrmzxGDBeW7zpo60IHzcfvg3rnD5+I13n8FMd5Dm/E
-8MDgJiw2r6Ul4aT+9WlqjmAw5Q0Jp+hdsNKyiYNsTaRFPYG+JYA8fdRSpqyxWY0C
-aYyKBUZXXWye+LIZHPEzAFgZyZArZrb2DmMh
------END CERTIFICATE-----
