+++ backend/pkg/manufacturing/mocks/client.go
@@ -1,25 +1,26 @@
 package mocks
 
 import (
+	"context"
 	"crypto"
 	"crypto/tls"
 	"crypto/x509"
 )
 
 type MockClient struct {
-	StartRemoteClientFn      func(CA string, authCRT []tls.Certificate) error
-	StartRemoteClientInvoked bool
+	StartClientFn      func(ctx context.Context, CA string, authCRT []tls.Certificate) error
+	StartClientInvoked bool
 
-	GetCertificateFn      func(keyAlg string, keySize int, c string, st string, l string, o string, ou string, cn string, email string) (*x509.Certificate, crypto.PrivateKey, error)
+	GetCertificateFn      func(ctx context.Context, keyAlg string, keySize int, c string, st string, l string, o string, ou string, cn string, email string) (*x509.Certificate, crypto.PrivateKey, error)
 	GetCertificateInvoked bool
 }
 
-func (mc *MockClient) StartRemoteClient(CA string, authCRT []tls.Certificate) error {
-	mc.StartRemoteClientInvoked = true
-	return mc.StartRemoteClientFn(CA, authCRT)
+func (mc *MockClient) StartClient(ctx context.Context, CA string, authCRT []tls.Certificate) error {
+	mc.StartClientInvoked = true
+	return mc.StartClientFn(ctx, CA, authCRT)
 }
 
-func (mc *MockClient) GetCertificate(keyAlg string, keySize int, c string, st string, l string, o string, ou string, cn string, email string) (*x509.Certificate, crypto.PrivateKey, error) {
+func (mc *MockClient) GetCertificate(ctx context.Context, keyAlg string, keySize int, c string, st string, l string, o string, ou string, cn string, email string) (*x509.Certificate, crypto.PrivateKey, error) {
 	mc.GetCertificateInvoked = true
-	return mc.GetCertificateFn(keyAlg, keySize, c, st, l, o, ou, cn, email)
+	return mc.GetCertificateFn(ctx, keyAlg, keySize, c, st, l, o, ou, cn, email)
 }
